pyproject.toml
[project]
name = "crypto-checker"
version = "0.1.0"
description = ""
authors = [
    {name = "5kiran",email = "dhrlfghks01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "dependency-injector (>=4.45.0,<5.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "crypto_checker", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
